+++ Flowering/DataModel/Plant+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  Plant+CoreDataProperties.swift
 //  Flowering
 //
-//  Created by Wenchu Du on 8/11/19.
+//  Created by Wenchu Du on 11/11/19.
 //  Copyright © 2019 Personal. All rights reserved.
 //
 //
@@ -17,11 +17,11 @@ extension Plant {
         return NSFetchRequest<Plant>(entityName: "Plant")
     }
 
-    @NSManaged public var fertility: Int32
+    @NSManaged public var fertility: Int16
     @NSManaged public var image: String?
-    @NSManaged public var moisture: Int32
+    @NSManaged public var moisture: Int16
     @NSManaged public var name: String?
-    @NSManaged public var sunshine: Int32
-    @NSManaged public var temperature: Int32
+    @NSManaged public var sunshine: Int16
+    @NSManaged public var temperature: Int16
 
 }
